SAK-4463
added some more configuration to the portal login/logout.  The following sakai.properties values are recognized:
login.text = the text you want for the login link or popup text for the login icon
login.icon = an image link for a graphic for the login to use instead of a link
xlogin.enabled = true | false - set to true to add the "xlogin" option, which is a bypass for container login that does an internal Sakai login
xlogin.text = the xlogin link or popup text wording
xlogin.icon = an image link for a graphic for xlogin
logout.text = test for the logout link or popup text for a logout icon
logout.icon = an image link for a graphic for logout to use instead of a link
login.url = an alternate URL that will be encoded to the login link or button

Note on login.url - you can send the users to an alternate login location, which in turn can login the user to Sakai via a web services call
then send them back with a URL to Sakai that encodes the user's session from login.

Note: also see change in r8459



git-svn-id: https://source.sakaiproject.org/svn/portal/trunk@8516 66ffb92e-73f9-0310-93c1-f5514f145a0a