Implement permission based button display.  This introduces the notion of required functions for a tool.
Each tool in its registration file indicates the Sakai functions which are needed for 
the execution of this tool.   For example, the Presentation tool is completely 
non-functional without content.read so the following line is added to the tool
registration:

    <configuration name="functions.require" value="content.read" />

This is a solution to SAK-4120.  This solution is different than proposed in SAK-4120 because
it is based on Sakai security functions rather than roles.  SAK-4120 also proposed an 
allow/deny pattern.  This may be evolved, but for now it is a pretty clean and simple 
notion that works for all site types and roles.  You can specify more than one security function
using commas.  The implication is "and" so if two functions are specfied the user must 
have *both* permissions.

SAK-4120



git-svn-id: https://source.sakaiproject.org/svn/portal/trunk@14751 66ffb92e-73f9-0310-93c1-f5514f145a0a