SAK-8894 SAK-9072 Fixed
SAK-8894 was not just when autoreset was being used the direct tool functionallity was completely broken, reported
seperately in SAK-9072. This was probably cause by the iFrame based reconstruct mechamism and the StoredState being 
reset without checking the contextpath of the destination of the stored state. I havent full tested the Iframe based re-construct 
since I dont have a suitable URL to test with. (To hand) Although notn of the code in this patch hits that sepcific area and
is only concerned with preventing the deletion of stored state for the wrong destination contexts.




git-svn-id: https://source.sakaiproject.org/svn/portal/trunk@27996 66ffb92e-73f9-0310-93c1-f5514f145a0a