SAK-11522
Make sure that the IMS Enterprise roles of "Admin" "Instructor" and "Student" work 
properly for all placements and all sites when the Servlet or Portlet isUserInRole 
is used.  This provides a simple compatibility for portable Portlets or Servlets
which need to run both within and outside of the Sakai Framework.



git-svn-id: https://source.sakaiproject.org/svn/portal/trunk@35133 66ffb92e-73f9-0310-93c1-f5514f145a0a