Basic Navigation worked out.
If you navigate to the site from another site and the first tool requests maximize 
you see maximize.

If there is only one tool on a site and it requests maximize, it always gets it.  In this
case the tool name is not even shown in the frame view - the site link becomes the "reset"
for the tool.

If in Maximize mode, you click on the site or my workspace - maximize mode is suppressed
for the next display - unless there is only one tool - you always see the tool list.  When
you then click on a tool - if it wants maximize it gets it.  This way even when all tools
want maximize, you can move between tools by clicking on the site link in maximize view.

The best usability will likely be if the first tool in the site (like the Home tool) does 
not request Maximize.  However - if this is the case, here is the sequence.  (1) Navigate 
to the site for the first time - first tool will be selected and shown maximally. (2) 
Click the Site tab - tools will be shown - the first tool will be selected and shown 
in non-maximized mode.  (3) Since you cannot click on a selected tool - you must click
on another tool and then click on the first tool to re-maximize the first tool, (4) 
if you click on a tool which also requests maximize, you must then click the site in
Maximize view and then click on the first tool.  The simple rule is that when you 
are in Maximized view and click on a site - you will see the tools unless there is only
one tool available.

Again - the best usability will happen if you build sites such that the first tool doest
not request Maximize unless there is only one tool on the site.



git-svn-id: https://source.sakaiproject.org/svn/portal/branches/SAK-12402@39643 66ffb92e-73f9-0310-93c1-f5514f145a0a