SAK-12402 SAK-12350
Clean up some cruft from SAK-12402.  As Ian refactored the code - he saw something 
that did not make sense - so he flagged it with a TODO.  This commit removes the cruft 
and addresses the TODO.

Ian Says:
#if ( $singleToolMap && // handled 
 $sakaiFrameSetRequested && // handled (will be false if above 
 $tabsSites && // this is true if there is a site view object 
( ! $sakaiFrameSuppress || // 
$sitePages.pageNavToolsCount == 1 ) ) // ????? this does not make any sense. 
 // it implies that we can ignore sakaiFrameSupress is pageNavToolsCount == 1,
 which will be true if singleToomMap == 1 hence sakaiFrameRequested = true

Chuck Says:
Nice catch.   Checking for pageNavToolsCount == 1 makes no sense.
It is cruft from an early version of SAK-12402 which took a different
approach to deciding when to switch to frame set mode.

The code in trunk only worked because (as you note above) the 
existence or lack of singleToolMap shields the case where
pageNavToolsCount would have been misleading.  I removed references
to sitePages and pageNavToolsCount from the frametop and frameset 
view code as it was neither used nor needed.



git-svn-id: https://source.sakaiproject.org/svn/portal/trunk@40420 66ffb92e-73f9-0310-93c1-f5514f145a0a