SAK-14262
Since JSR-168 portlets do not use iframes - they are included in the background document in 
the /portal/page output. Normal tools have an iframe tag in the /portal/page output and 
the src= for the iframs references the /portal/tool/toolid to get the tool markup. 

Normally there is no problem unless you make a URL that contains the toolid of 
a JSR-168 tool and use it from outside the regular Sakai portal.  

This fix makes the /portal/tool/toolid URLs work in a decent compatibility mode 
in case these are used.

This is likely back-portable to 2-5-x but since there have been no reports of this,
it is likely not necessary.  The key is to prepare for JSR-168 support in the new
flat portal.



git-svn-id: https://source.sakaiproject.org/svn/portal/trunk@51250 66ffb92e-73f9-0310-93c1-f5514f145a0a