SAK-21626 BLTI-76
Make it possible for different Basic LTI tools to have different 
icons - particular when new tools are pre-registered. This feature 
actually makes is possible so that all tools can choose their own 
icon class. This is done by adding a property of 

sakai:menuClass 

To either a Tool or a Placement. This overrides the default class 
based on the tool registration id like: 

icon-sakai-lessonbuildertool 

It still requires the icon image to be added and referenced in 
a CSS class - but since this is done with CSS there is 
no real way to do a better job of short-circuting this.



git-svn-id: https://source.sakaiproject.org/svn/portal/trunk@102497 66ffb92e-73f9-0310-93c1-f5514f145a0a