SAK-22071 SAK-22199
More progress and increasing sophistication in the PDA portal's ability to inline
Wicket-based applications.  The PDA portal will not inline if 

o The 'wicket-ajax' header is set in the incoming request
o The URI ende in '.prf' 
o The query string matches the regex: 'wicket:interface=.*iresourcelistener:' (lowercased)

This is a bit tricky because the mere presense of a wicket:interface= is not sufficiently
accurate by itself.

There may yet be patterns we have not caught.  But now profile2 and roster seem to pass 
a quick run-through QA with all images and AJAX bits working as well as roster download.

Again a big thanks to Yuki for testing all these and catching all the various combinations.



git-svn-id: https://source.sakaiproject.org/svn/portal/trunk@108569 66ffb92e-73f9-0310-93c1-f5514f145a0a