SAK-22633
Makes it so the *only* URLs that should be forwarded to PDAHandler are 

/portal, /portal/pda/... and /portal/site/... 

The other URLs need to be handled normally by which ever handler is appropriate. 
In particular /portal/tool should *never* be sent to PDAHandler as the PDA output 
when using iframes used /portal/tool/... to display the within frame content. 



git-svn-id: https://source.sakaiproject.org/svn/portal/trunk@113107 66ffb92e-73f9-0310-93c1-f5514f145a0a