SAK-25503
Refactor some of the logic into a utility class.   Add code to catch urls of the form:

http://localhost:8080/portal/site/4f5ea191/page/41a8c075

That should really be 

http://localhost:8080/portal/site/4f5ea191/tool/12345

And redirect to the apporpriate URL to do the inlining. 

This catches the site to site navigation, tab drop-down, drawer drop down, 
little green person drop down, and bookmarked URLs.



git-svn-id: https://source.sakaiproject.org/svn/portal/trunk@132914 66ffb92e-73f9-0310-93c1-f5514f145a0a